+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.9.0
+
+* Add solver configuration support
+* Add optimization: Minimum Domain Size
+* Add optimization: Maximum Degree 
+
 ## 0.8.0
 
 * Use crypto.random for random generation